les: use timeutil
